LUCENE-3220: fix integer math division bug in I(ne) model
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/flexscoring@1155434 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/test/org/apache/lucene/search/similarities/BasicModelIne.java
@@ -29,7 +29,7 @@ public class BasicModelIne extends BasicModel {
   public final float score(EasyStats stats, float tfn) {
     int N = stats.getNumberOfDocuments();
     long F = stats.getTotalTermFreq();
-    double ne = N * (1 - Math.pow((N - 1) / N, F));
+    double ne = N * (1 - Math.pow((N - 1) / (double)N, F));
     return tfn * (float)(log2((N + 1) / (ne + 0.5)));
   }
 }
